teams_transcript_ppt_simple/requirements.txt
fastapi
uvicorn
requests
python-pptx
python-multipart
openai
python-docx